+++ internal/pkg/request/url-mapper.go
@@ -12,39 +12,6 @@ import (
 	"strings"
 )
 
-type PathAmbiguousError struct {
-	RequestUrl string
-	FirstMatch string
-	OtherMatch string
-}
-
-type PathNotFoundError struct {
-	RequestUrl string
-}
-
-type MapperOutput struct {
-	Datastore string
-	Package   string
-}
-
-func (e *PathAmbiguousError) Error() string {
-	return fmt.Sprintf("Path-mapping [%s] is ambiguous! Mapping [%s] also matches incoming path [%s]!", e.RequestUrl, e.FirstMatch, e.OtherMatch)
-}
-
-func (e *PathNotFoundError) Error() string {
-	return fmt.Sprintf("PathMapper: There is no mapping which matches path [%s]!", e.RequestUrl)
-}
-
-type PathMapper interface {
-	Configure(appConf *configs.AppConfig) error
-	Map(interface{}) (*MapperOutput, error)
-}
-
-type pathMapperInput struct {
-	Method string
-	Url    *url.URL
-}
-
 type pathMapper struct {
 	appConf    *configs.AppConfig
 	mappings   []*compiledMapping
@@ -56,6 +23,11 @@ type compiledMapping struct {
 	mapping    *configs.ApiMapping
 	importance int
 	datastore  string
+}
+
+type pathMapperInput struct {
+	Method string
+	Url    *url.URL
 }
 
 func NewPathMapper() PathMapper {
