+++ internal/pkg/request/url-processor.go
@@ -9,22 +9,6 @@ import (
 	"strings"
 )
 
-type PathProcessorConfig struct {
-	PathMapper *PathMapper
-}
-
-type PathProcessorOutput struct {
-	Datastore string
-	Package   string
-	Path      []string
-	Queries   map[string]interface{}
-}
-
-type PathProcessor interface {
-	Configure(appConf *configs.AppConfig, processorConf *PathProcessorConfig) error
-	Process(input interface{}) (*PathProcessorOutput, error)
-}
-
 type urlProcessor struct {
 	appConf    *configs.AppConfig
 	config     *PathProcessorConfig
